+++ inputremapper/configs/validation_errors.py
@@ -117,3 +117,16 @@ class MacroParsingError(ValueError):
     def __init__(self, symbol: Optional[str] = None, msg="Error while parsing a macro"):
         self.symbol = symbol
         super().__init__(msg)
+
+
+def pydantify(error: type):
+    """Generate a string as it would appear IN pydantic error types.
+
+    This does not include the base class name, which is transformed to snake case in
+    pydantic. Example pydantic error type: "value_error.foobar" for FooBarError.
+    """
+    # See https://github.com/pydantic/pydantic/discussions/5112
+    lower_classname = error.__name__.lower()
+    if lower_classname.endswith("error"):
+        return lower_classname[: -len("error")]
+    return lower_classname
